Refactor unzip to non-R file
This is needed for the unzip process, to call it
from another binary.
