+++ src/Makevars
@@ -1,11 +1,11 @@
 
-OBJECTS = init.o miniz.o rzip.o
+OBJECTS = init.o miniz.o rzip.o zip.o
 
 .PHONY: all clean
 
 all: tools/unzip $(SHLIB)
 
-tools/unzip: miniz.c tools/unzip.c
+tools/unzip: miniz.c zip.c tools/unzip.c
 	$(CC) $(CFLAGS) miniz.c tools/unzip.c -o tools/unzip
 
 clean:
