+++ src/Makevars.win
@@ -1,11 +1,11 @@
 
-OBJECTS = init.o miniz.o rzip.o
+OBJECTS = init.o miniz.o rzip.o zip.o
 
 .PHONY: all clean
 
 all: tools/unzip $(SHLIB)
 
-tools/unzip.exe: miniz.c tools/unzip.c
+tools/unzip.exe: miniz.c zip.c tools/unzip.c
 	$(CC) $(CFLAGS) miniz.c tools/unzip.c -o tools/unzip
 
 clean:
